requirements.txt
aiogram==3.15.0
python-dotenv==1.0.1
aiosqlite==0.20.0
groq>=1.0.0
